+++ Source/Runtime/Executor/Mono/ButtonCommandExecutor.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace LazerLabs.Commands
+{
+    public abstract class ButtonCommandExecutor : BaseButtonExecutor<Action>
+    {
+        protected abstract ICommand Command { get; }
+        protected override Action Context => Command.Execute;
+    }
+}