+++ Source/Runtime/Executor/Mono/Core/MonoExecutor.cs
@@ -1,15 +1,16 @@
 using System;
-using System.Threading.Tasks;
 using UnityEngine;
 
 namespace LazerLabs.Commands
 {
-    public sealed class MonoButtonExecutor : ButtonCommandExecutor
+    public abstract class MonoExecutor : CommandMonoExecutor
     {
         [SerializeField] private CommandMonoRunner m_commandRunner = default;
         [SerializeField] private MonoCommand m_monoCommand = default;
 
+
         protected override ICommandVoid<Action> Runner => m_commandRunner;
         protected override ICommand Command => m_monoCommand;
+
     }
 }