+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.0.6] - 2024-04-10
+### Added
+- Async runner and executor
+### Changed
+- Rework executors
+
+
 ## [0.0.5] - 2024-04-06
 ### Added
 - Added mono other one input Executor